pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skcomm"
version = "0.1.0"
description = "Transport-agnostic, redundant, encrypted communication for sovereign AI agents"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    {name = "smilinTux", email = "admin@smilintux.org"},
]
keywords = [
    "ai", "agent", "communication", "p2p", "encrypted",
    "transport", "messaging", "sovereign", "syncthing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Communications",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cli = [
    "click>=8.1",
    "rich>=13.0",
]
crypto = [
    "capauth>=0.1.0",
]
nostr = [
    "websockets>=12.0",
    "cryptography>=42.0",
]
discovery = [
    "zeroconf>=0.131.0",
]
all = [
    "click>=8.1",
    "rich>=13.0",
    "capauth>=0.1.0",
    "websockets>=12.0",
    "cryptography>=42.0",
    "zeroconf>=0.131.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.3",
]

[project.scripts]
skcomm = "skcomm.cli:main"

[project.urls]
Homepage = "https://smilintux.org"
Repository = "https://github.com/smilinTux/skcomm"
Issues = "https://github.com/smilinTux/skcomm/issues"
Changelog = "https://github.com/smilinTux/skcomm/releases"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 99
target-version = ["py310"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
